补充webapi

+++ pkg/servers/material/logic/material_tray.go
@@ -1,18 +1,36 @@
 package logic
 
 import (
+	"fmt"
+
 	"github.com/CloudSilk/CloudSilk/pkg/model"
 	"github.com/CloudSilk/CloudSilk/pkg/proto"
+	"github.com/CloudSilk/CloudSilk/pkg/types"
 	"github.com/CloudSilk/pkg/utils"
 	"gorm.io/gorm/clause"
 )
 
 func CreateMaterialTray(m *model.MaterialTray) (string, error) {
+	m.CurrentState = types.MaterialTrayStateWaitFill
+	if !m.Enable {
+		m.CurrentState = types.MaterialTrayStateDisabled
+	}
 	err := model.DB.DB().Create(m).Error
 	return m.ID, err
 }
 
 func UpdateMaterialTray(m *model.MaterialTray) error {
+	var materialTray model.MaterialTray
+	if err := model.DB.DB().Preload("ProductInfo").First(&materialTray, "`id` = ?", m.ID).Error; err != nil {
+		return err
+	}
+	if !m.Enable && materialTray.ProductInfoID != nil {
+		return fmt.Errorf("停用失败，当前托盘已绑定序列号为%s的产品。", materialTray.ProductInfo.ProductSerialNo)
+	}
+	if m.ProductionLineID != materialTray.ProductionLineID && materialTray.ProductInfoID != nil {
+		return fmt.Errorf("变更产线失败，当前托盘已绑定序列号为%s的产品。", materialTray.ProductInfo.ProductSerialNo)
+	}
+
 	return model.DB.DB().Omit("created_at").Save(m).Error
 }
 
@@ -69,5 +87,13 @@ func GetMaterialTrayByIDs(ids []string) ([]*model.MaterialTray, error) {
 }
 
 func DeleteMaterialTray(id string) (err error) {
+	var productInfoID string
+	if err := model.DB.DB().Model(&model.MaterialTray{}).Where("`id` = ?", id).Select("product_info_id").Scan(&productInfoID).Error; err != nil {
+		return err
+	}
+	if productInfoID != "" {
+		return fmt.Errorf("托盘正在使用中，无法删除")
+	}
+
 	return model.DB.DB().Delete(&model.MaterialTray{}, "`id` = ?", id).Error
 }
